HomeFragemnet:
We have 2 threads running in parallel in HomeFragment. 1st thread is responsible for keeping the client server connection...
2nd thread is for fetching data and displaying it...

UserProfile: UserProfile is better view and fixed the username saving issue with the view and better implementatio+ ongridListener
MongoDbSetup: changed few implementations in it, like RunAppClient is done once inside the returned instance.

server client working and exchanging data repeatedly after the time set for it.
runs on separate thread to avoid Exception (onMainThreadNetwork error).